+++ changelog.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.7] - 2026-07-15
 
 ### Changed
 
@@ -111,8 +111,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - Previous Release
 
-[Unreleased]: https://github.com/rockingrow/algo-trading-worker/compare/v1.1.6...dev
-[1.1.6]: https://github.com/rockingrow/algo-trading-worker/compare/v1.1.5...dev
+[1.1.7]: https://github.com/rockingrow/algo-trading-worker/compare/v1.1.7...dev
+[1.1.6]: https://github.com/rockingrow/algo-trading-worker/compare/v1.1.6...dev
 [1.1.5]: https://github.com/rockingrow/algo-trading-worker/compare/v1.1.4...v1.1.5
 [1.1.4]: https://github.com/rockingrow/algo-trading-worker/compare/v1.1.3...dev
 [1.1.3]: https://github.com/rockingrow/algo-trading-worker/compare/v1.1.2...v1.1.3